requirements.txt
streamlit==1.37.1
requests==2.32.3
python-dotenv==1.0.1

langchain-core==0.3.29
langchain-groq==0.2.1
langchain-huggingface==0.1.2

sentence-transformers==3.0.1
torch>=2.1.0

pdfplumber==0.11.4
python-docx==1.1.2
